move sqlhelper into util/. Add convenience methods
