+++ test/tpc_h_tests.cpp
@@ -1,7 +1,9 @@
 #include "thirdparty/microtest/microtest.h"
-#include "sql_asserts.h"
+
 #include "SQLParser.h"
-#include "sqlhelper.h"
+#include "util/sqlhelper.h"
+
+#include "sql_asserts.h"
 
 #include <string>
 #include <fstream>
